Move examples importer widget to app utils
